refactor: move get_proxied_source to RegistryClient to avoid helpers class
