+++ terraform/lib/dependabot/terraform/registry_client.rb
@@ -3,7 +3,6 @@
 require "dependabot/dependency"
 require "dependabot/errors"
 require "dependabot/source"
-require "dependabot/terraform/helpers"
 require "dependabot/terraform/version"
 
 module Dependabot
@@ -11,6 +10,7 @@ module Dependabot
     # Terraform::RegistryClient is a basic API client to interact with a
     # terraform registry: https://www.terraform.io/docs/registry/api.html
     class RegistryClient
+      ARCHIVE_EXTENSIONS = %w(.zip .tbz2 .tgz .txz).freeze
       PUBLIC_HOSTNAME = "registry.terraform.io"
 
       def initialize(hostname: PUBLIC_HOSTNAME, credentials: [])
@@ -19,6 +19,39 @@ module Dependabot
           memo[item["host"]] = item["token"] if item["type"] == "terraform_registry"
         end
       end
+
+      # rubocop:disable Metrics/PerceivedComplexity
+      # See https://www.terraform.io/docs/modules/sources.html#http-urls for
+      # details of how Terraform handle HTTP(S) sources for modules
+      def self.get_proxied_source(raw_source) # rubocop:disable Metrics/AbcSize
+        return raw_source unless raw_source.start_with?("http")
+
+        uri = URI.parse(raw_source.split(%r{(?<!:)//}).first)
+        return raw_source if uri.path.end_with?(*ARCHIVE_EXTENSIONS)
+        return raw_source if URI.parse(raw_source).query&.include?("archive=")
+
+        url = raw_source.split(%r{(?<!:)//}).first + "?terraform-get=1"
+        host = URI.parse(raw_source).host
+
+        response = Excon.get(
+          url,
+          idempotent: true,
+          **SharedHelpers.excon_defaults
+        )
+        raise PrivateSourceAuthenticationFailure, host if response.status == 401
+
+        return response.headers["X-Terraform-Get"] if response.headers["X-Terraform-Get"]
+
+        doc = Nokogiri::XML(response.body)
+        doc.css("meta").find do |tag|
+          tag.attributes&.fetch("name", nil)&.value == "terraform-get"
+        end&.attributes&.fetch("content", nil)&.value
+      rescue Excon::Error::Socket, Excon::Error::Timeout => e
+        raise PrivateSourceAuthenticationFailure, host if e.message.include?("no address for")
+
+        raw_source
+      end
+      # rubocop:enable Metrics/PerceivedComplexity
 
       # Fetch all the versions of a provider, and return a Version
       # representation of them.
@@ -77,7 +110,7 @@ module Dependabot
             source_url.start_with?("/") ||
             source_url.start_with?("./") ||
             source_url.start_with?("../")
-          source_url = Helpers.get_proxied_source(source_url) if source_url
+          source_url = RegistryClient.get_proxied_source(source_url) if source_url
         when "provider", "providers"
           response = http_get(URI.join(base_url, "#{dependency.name}/#{dependency.version}"))
           return nil unless response.status == 200
